This is a replacement for the current script file..
This takes into account that the value of the const var is empty and only a refresh of the browser will make the values stick. 
So i utilized a button to update the values of the now variable vars and send the user-data and return nothing after the button has been clicked..
Much better than what we started with.