This file compiles into addInt.wasm
Use wat-wasm,to compile to wasm      --       available via:      npm install wat-wasm -g
This file is a representation in S-expressions of code written by Rick Battagline in 'The Art of WebAssembly...'
The file allows the adding of 2 integers that are passed as arguments to the exported function. 
Full code can be found for wat-wasm https://github.com/battlelinegames
and extensive code can be found in The Art of WebAssembly: Builld Secure, Portable, High-Performance Applications.